Mostly debugged an issue with sorting in --nofilter mode.
Fixed an issue with the default value for the vcf outfile suffix default.
Removed --input-file as a flag alias.
Changed the usage description of --filter|--no-filter to work in either case.
Changed the initial value of best_score to -1 so that cases where there was no data get sorted below cases where there was data but no evidence.
Changed the initial value of best_pair to '.' so that the column value wouldn't be an empty string when there was no data.
Added a super-sort on no-data versus data so that no-data cases would go to the bottom in the --no-filter mode.
Added a way to equate decimal values so that the sub-sort on read depth would work.
Changed getBestScoringData to better handle cases of no data when --no-filter is true.
Removed the vestigial rank subroutine.

Files checked in: src/rankVCF.pl