Added tests for structural variant data.
I also had to fix the code involved in checking exact file matches. The file contents of files that are undefined needed to also be undefined. The check of whether file contents should be checked was being done on the outfile instead of on the "shoulds" array. I also had to change the check for diff failure to require that the stdout from it was empty.

Files checked in: t/run_tests.t