Closed STDIN at the top of the script.
Conda was passing an active STDIN handle as a PIPE to this test script, which was then getting inherited by calls to vcfSampleCompare.pl inside backticks, which proceeded to wait forever for input. Conda should probably NOT be piping to the test script, but regardless, closing STDIN in the test script seems prudent to avoid unexpected empty input.

Files checked in: t/run_tests.t