Consumed a possible stdin file if standard in is passed to the test script.
This was an issue when running the test script during a conda build, which was giving the script its standard input by inheritance. Closing STDIN worked to allow the tests to run without inheriting the standard input, but the test script was issuing a warning on exit that not all input file sets had been processed. The input file type was only created to hide it in the first place, so I added a call to getInfile() to suppress the warning.

What really needs to be done is allow CLI to run in a mode where it takes no input files.

Files checked in: t/run_tests.t