Finished edge scoring (issue #9), orig sample order (issue #14), and fixed a bug that should have affected the growing of sample groups.
The bug was undocumented.  I'm not 100ure it was a bug.  It could have just been indecipherable logic, but it looked like growing sample groups was adding the next sample that resulted in the smallest gap, not the largest.  It only affected groups based on observation frequency and my testing showed no difference.  The replacement code, at the very least, is easier to understand.

The script still defaults to measuring sample group difference using the mean, but edge gap scoring is now possible.

In order to be better able to compare outputs using the different scoring measures, I decided to sort the samples in each group based on the original order in the input file in the headers.  And in the case of dynamically build sample groups, I also order the groups themselves (in a pair) based on the first sample in each group and the original order.

Files checked in: src/vcfSampleCompare.pl