Addressed requirement #10 - ratio of resolution.
Added a scoring function based on ratio of resolution.
Updated the GT_SCOREs to report ratio of resolution.
Added an advanced option to set a threshold for ratio of resolution.
Changed the defaults for depth.
Changed the default for -a.
Changed the behavior of growth of sample groups when not filtering.  Filtered sample groups will now grow as long as growth improves or maintains their scores.

Files checked in: src/vcfSampleCompare.pl